Add ImGui

+++ Source/Position.cpp
@@ -4,13 +4,13 @@
 
 namespace SpaceSim
 {
-    double DistanceBetween(const Position &a, const Position &b)
+    float DistanceBetween(const Position &a, const Position &b)
     {
-        return std::sqrt(std::pow(a.X + b.X, 2) + std::pow(a.Y + b.Y, 2));
+        return std::sqrtf(std::powf(a.X + b.X, 2) + std::powf(a.Y + b.Y, 2));
     }
 
-    double AngleBetween(const Position &a, const Position &b)
+    float AngleBetween(const Position &a, const Position &b)
     {
-        return std::atan2(a.Y - b.Y, a.X - b.X);
+        return std::atan2f(a.Y - b.Y, a.X - b.X);
     }
 } // namespace SpaceSim
